# [metrics] Fix perf regression introduced by metric point reclaim feature