--- a/src/OpenTelemetry/Metrics/AggregatorStore.cs
+++ b/src/OpenTelemetry/Metrics/AggregatorStore.cs
@@ -12,7 +12,7 @@ namespace OpenTelemetry.Metrics;
 internal sealed class AggregatorStore
 {
     internal readonly bool OutputDelta;
-    internal readonly bool ShouldReclaimUnusedMetricPoints;
+    internal readonly bool OutputDeltaWithUnusedMetricPointReclaimEnabled;
     internal long DroppedMeasurements = 0;
 
     private static readonly string MetricPointCapHitFixMessage = "Consider opting in for the experimental SDK feature to emit all the throttled metrics under the overflow attribute by setting env variable OTEL_DOTNET_EXPERIMENTAL_METRICS_EMIT_OVERFLOW_ATTRIBUTE = true. You could also modify instrumentation to reduce the number of unique key/value pair combinations. Or use Views to drop unwanted tags. Or use MeterProviderBuilder.SetMaxMetricPointsPerMetricStream to set higher limit.";
@@ -101,9 +101,9 @@ internal sealed class AggregatorStore
             reservedMetricPointsCount++;
         }
 
-        this.ShouldReclaimUnusedMetricPoints = shouldReclaimUnusedMetricPoints;
+        this.OutputDeltaWithUnusedMetricPointReclaimEnabled = shouldReclaimUnusedMetricPoints && this.OutputDelta;
 
-        if (this.OutputDelta && shouldReclaimUnusedMetricPoints)
+        if (this.OutputDeltaWithUnusedMetricPointReclaimEnabled)
         {
             this.availableMetricPoints = new Queue<int>(maxMetricPoints - reservedMetricPointsCount);
 
@@ -158,17 +158,14 @@ internal sealed class AggregatorStore
     internal int Snapshot()
     {
         this.batchSize = 0;
-        if (this.OutputDelta)
+        if (this.OutputDeltaWithUnusedMetricPointReclaimEnabled)
         {
-            if (this.ShouldReclaimUnusedMetricPoints)
-            {
-                this.SnapshotDeltaWithMetricPointReclaim();
-            }
-            else
-            {
-                var indexSnapshot = Math.Min(this.metricPointIndex, this.maxMetricPoints - 1);
-                this.SnapshotDelta(indexSnapshot);
-            }
+            this.SnapshotDeltaWithMetricPointReclaim();
+        }
+        else if (this.OutputDelta)
+        {
+            var indexSnapshot = Math.Min(this.metricPointIndex, this.maxMetricPoints - 1);
+            this.SnapshotDelta(indexSnapshot);
         }
         else
         {