[Noticed this while working on #5131]

## Changes

* Fixes a perf regression introduced by metric point reclaim for delta temporality even when disabled.

## Benchmarks

Before:
| Method         | AggregationTemporality | Mean     | Error    | StdDev   |
|--------------- |----------------------- |---------:|---------:|---------:|
| CounterHotPath | Cumulative             | 10.07 ns | 0.058 ns | 0.054 ns |
| CounterHotPath | Delta                  | 13.60 ns | 0.179 ns | 0.159 ns |

After:
| Method         | AggregationTemporality | Mean     | Error    | StdDev   |
|--------------- |----------------------- |---------:|---------:|---------:|
| CounterHotPath | Cumulative             | 10.93 ns | 0.185 ns | 0.155 ns |
| CounterHotPath | Delta                  | **10.17 ns** | 0.117 ns | 0.103 ns |

## Merge requirement checklist

* [X] [CONTRIBUTING](https://github.com/open-telemetry/opentelemetry-dotnet/blob/main/CONTRIBUTING.md) guidelines followed (license requirements, nullable enabled, static analysis, etc.)